// src/shared/api/generated/models/assignedVacancyRecruiterDto.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * VIO Stats Service API
 * API для управления компаниями, рекрутерами и вакансиями
 * OpenAPI spec version: 1.0
 */
import type { AssignedRecruiterDto } from './assignedRecruiterDto';
import type { AssignedVacancyDto } from './assignedVacancyDto';

export interface AssignedVacancyRecruiterDto {
  /** Уникальный идентификатор назначения */
  id: string;
  /** ID рекрутера */
  recruiterId: string;
  /** ID вакансии */
  vacancyId: string;
  /** Количество необходимых резюме */
  requiredResumes: number;
  /** Количество отправленных резюме */
  sentResumes: number;
  /** Количество принятых резюме */
  acceptedResumes: number;
  /** Количество отклоненных резюме */
  rejectedResumes: number;
  /** Дата создания */
  createdAt: string;
  /** Дата обновления */
  updatedAt: string;
  recruiter: AssignedRecruiterDto;
  vacancy: AssignedVacancyDto;
}